try to remove runtime err
