+++ src/vs/workbench/contrib/search/browser/searchTreeModel/folderMatch.ts
@@ -18,7 +18,7 @@ import { IFileMatch, IPatternInfo, ITextQuery, ITextSearchPreviewOptions, result
 
 import { FileMatchImpl } from './fileMatch.js';
 import { IChangeEvent } from './searchTreeCommon.js';
-import { IFileInstanceMatch, IFolderMatch, IFolderMatchWithResource, IFolderMatchNoRoot as IFolderMatchNoRoot, IFolderMatchWorkspaceRoot, ISearchModel, ISearchResult, isFileInstanceMatch, isFolderMatch, isFolderMatchWorkspaceRoot, ITextSearchHeading, isFolderMatchNoRoot } from './searchTreeCommon.js';
+import { IFileInstanceMatch, IFolderMatch, IFolderMatchWithResource, IFolderMatchNoRoot, IFolderMatchWorkspaceRoot, ISearchModel, ISearchResult, isFileInstanceMatch, isFolderMatch, isFolderMatchWorkspaceRoot, ITextSearchHeading, isFolderMatchNoRoot } from './searchTreeCommon.js';
 import { NotebookEditorWidget } from '../../../notebook/browser/notebookEditorWidget.js';
 import { isINotebookFileMatchNoModel } from '../../common/searchNotebookHelpers.js';
 import { NotebookCompatibleFileMatch } from '../notebookSearch/notebookSearchModel.js';
